I finished the task at hand

+++ src/shapes/Rectangle.java
@@ -5,23 +5,23 @@ public class Rectangle extends Quadrilateral implements Measurable{
     public Rectangle(double length, double width){
         super(length, width);
     }
-////
+
     @Override
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
     @Override
-    public void setLength(int length) {
+    public void setLength(double length) {
         this.length = length;
     }
     @Override
-    public int getArea(){
-        int area = this.length * this.width;
+    public double getArea(){
+        double area = this.length * this.width;
         return area;
     }
     @Override
-    public int getPerimeter(){
-        int perimeter = 2 * this.length + 2 * this.width;
+    public double getPerimeter(){
+        double perimeter = 2 * this.length + 2 * this.width;
         return perimeter;
     }
